feat(backend): moved batch logic into batchSet object
